refactor: main electron with managers and handlers
