+++ electron/package.json
@@ -67,6 +67,7 @@
   },
   "dependencies": {
     "@npmcli/arborist": "^7.1.0",
+    "@types/request": "^2.48.12",
     "@uiball/loaders": "^1.3.0",
     "electron-store": "^8.1.0",
     "electron-updater": "^6.1.4",
