--- server/nodemon.json
@@ -1,5 +1,0 @@
-{
-    "watch": [
-        "main.ts"
-    ]
-}
