--- server/package.json
@@ -1,26 +1,0 @@
-{
-  "name": "server",
-  "version": "1.0.0",
-  "main": "index.js",
-  "license": "MIT",
-  "dependencies": {
-    "cors": "^2.8.5",
-    "electron": "^26.2.1",
-    "express": "^4.18.2",
-    "request": "^2.88.2",
-    "request-progress": "^3.0.0"
-  },
-  "devDependencies": {
-    "@types/cors": "^2.8.14",
-    "@types/express": "^4.17.18",
-    "@types/node": "^20.8.2",
-    "nodemon": "^3.0.1",
-    "ts-node": "^10.9.1",
-    "typescript": "^5.2.2"
-  },
-  "scripts": {
-    "build": "tsc --project ./",
-    "dev": "nodemon main.ts",
-    "prod": "node build/main.js"
-  }
-}
