--- server/tsconfig.json
@@ -1,19 +1,0 @@
-{
-    "compilerOptions": {
-      "target": "es5",
-      "module": "commonjs",
-      "noImplicitAny": true,
-      "sourceMap": true,
-      "strict": true,
-      "outDir": "./build",
-      "rootDir": "./",
-      "noEmitOnError": true,
-      "baseUrl": ".",
-      "allowJs": true,
-      "paths": { "*": ["node_modules/*"] },
-      "typeRoots": ["node_modules/@types"],
-      "esModuleInterop": true
-    },
-    "include": ["./**/*.ts"],
-    "exclude": ["core", "build", "dist", "tests"]
-  }
